+++ libraries/Utility/src/Tuple.cpp
@@ -8,10 +8,10 @@ namespace raytracer {
 namespace utility { 
 
 bool Tuple::isVector() const {
-  return floatNearlyEqual(this->w(), 0.0f);
+  return floatNearlyEqual(this->w(), 0.0);
 }
 bool Tuple::isPoint() const {
-  return floatNearlyEqual(this->w(), 1.0f);
+  return floatNearlyEqual(this->w(), 1.0);
 }
 
 bool Tuple::operator==(const Tuple& rhs) const {
@@ -33,11 +33,11 @@ Tuple Tuple::operator-(const Tuple& rhs)  const{
   return Tuple(this->x()-rhs.x(), this->y()-rhs.y(), 
                this->z()-rhs.z(), this->w()-rhs.w());
 }
-Tuple Tuple::operator*(const float& rhs) const{
+Tuple Tuple::operator*(const double& rhs) const{
   return Tuple(this->x()*rhs, this->y()*rhs, 
                this->z()*rhs, this->w()*rhs);
 }
-Tuple Tuple::operator/(const float& rhs) const{
+Tuple Tuple::operator/(const double& rhs) const{
   return Tuple(this->x()/rhs, this->y()/rhs, 
                this->z()/rhs, this->w()/rhs);
 }
@@ -45,35 +45,35 @@ Tuple Tuple::operator-() const{
   return *this*-1;
 }
 
-const float& Tuple::x() const{
+const double& Tuple::x() const{
   return data[0];
 }
-float& Tuple::x(){
-  return const_cast<float&>(const_cast<const Tuple*>(this)->data[0]);
+double& Tuple::x(){
+  return const_cast<double&>(const_cast<const Tuple*>(this)->data[0]);
 }
 
-const float& Tuple::y() const{
+const double& Tuple::y() const{
   return data[1];
 }
-float& Tuple::y(){
-  return const_cast<float&>(const_cast<const Tuple*>(this)->data[1]);
+double& Tuple::y(){
+  return const_cast<double&>(const_cast<const Tuple*>(this)->data[1]);
 }
 
-const float& Tuple::z() const{
+const double& Tuple::z() const{
   return data[2];
 }
-float& Tuple::z(){
-  return const_cast<float&>(const_cast<const Tuple*>(this)->data[2]);
+double& Tuple::z(){
+  return const_cast<double&>(const_cast<const Tuple*>(this)->data[2]);
 }
 
-const float& Tuple::w() const{
+const double& Tuple::w() const{
   return data[3];
 }
-float& Tuple::w(){
-  return const_cast<float&>(const_cast<const Tuple*>(this)->data[3]);
+double& Tuple::w(){
+  return const_cast<double&>(const_cast<const Tuple*>(this)->data[3]);
 }
 
-float Tuple::magnitude() const{
+double Tuple::magnitude() const{
   auto sumSquared = this->x()*this->x() + this->y()*this->y() +
                     this->z()*this->z() + this->w()*this->w();
   return std::sqrt(sumSquared);
@@ -83,7 +83,7 @@ Tuple Tuple::normalize() const{
   return *this/this->magnitude();
 }
 
-float Tuple::dot(const Tuple& rhs) const{
+double Tuple::dot(const Tuple& rhs) const{
   auto multAndACum = this->x()*rhs.x() + this->y()*rhs.y() +
                      this->z()*rhs.z() + this->w()*rhs.w();
   return multAndACum;
@@ -99,10 +99,10 @@ Tuple Tuple::reflect(const Tuple& normal) const{
   return *this - (normal * 2 * this->dot(normal));
 }
 
-Tuple Point(float x, float y, float z){
+Tuple Point(double x, double y, double z){
   return Tuple(x,y,z,1);
 }
-Tuple Vector(float x, float y, float z){
+Tuple Vector(double x, double y, double z){
   return Tuple(x,y,z,0);
 }
